# minicpm: make embedding_scale residual_scale logit_scale optional with legacy defaults.